+++ test/Utils.test.ts
@@ -1,5 +1,6 @@
 import { expect } from "chai";
-import { resolveType } from "../src/utils";
+import { isEqual, isObject } from "lodash";
+import { resolveType, isChanged, changes } from "../src/utils";
 
 describe("Model Utils", () => {
   it("should resolve primitive value types", () => {
@@ -7,5 +8,27 @@ describe("Model Utils", () => {
     expect(resolveType("john")).to.be.equal("String");
     expect(resolveType(true)).to.be.equal("Boolean");
     expect(resolveType({}, "userInput")).to.be.equal("UserInput");
+  });
+});
+
+describe("isChanged Utils", () => {
+  it("expects objects with array attr to be changed", () => {
+    const a = {
+      id: "1",
+      tags: ["A", "B"],
+    };
+    const b = {
+      id: "2",
+    };
+    const c = {
+      tags: ["A"],
+    };
+    const d = {
+      tags: ["A", "B", "C"],
+    };
+
+    expect(isChanged(a, b)).to.equal(true);
+    expect(isChanged(a, c)).to.equal(true);
+    expect(isChanged(a, d)).to.equal(true);
   });
 });